Driver Radi, Pravim APP

+++ aplikacija/app.c
@@ -130,7 +130,7 @@ label1:    printf("Unesite broj - clanova niza: ");
         goto label1;
     }
 
-    float tx_buffer[array_num];
+    uint tx_buffer[array_num];
     float rx_buffer[array_num];
     float rx_buffer_cpu[array_num];
 
@@ -150,7 +150,7 @@ label1:    printf("Unesite broj - clanova niza: ");
             printf("DEBUG --- VALUE je: %f \n", value);
             printf("DEBUG --- FloatToHex je: %#x \n", floatToHex(value));
             value = floatToHex(value);
-            tx_buffer[i] = (float)value;
+            tx_buffer[i] = (uint)value;
             printf("DEBUG --- NOVI VALUE je: %#x \n", (uint)value);
             printf("DEBUG --- NOVI VALUE je: %f \n", hexToFloat(value));
             printf("Na poziciji %d nalazi se vrednost: %#x\n", i, tx_buffer[i]);
